+++ src/basic/models/products/useProductForm.ts
@@ -1,0 +1,35 @@
+import { useState } from "react";
+import { ProductWithUI } from "./types";
+
+export const useProductForm = () => {
+  const [productForm, setProductForm] = useState({
+    name: "",
+    price: 0,
+    stock: 0,
+    description: "",
+    discounts: [] as Array<{ quantity: number; rate: number }>,
+  });
+  const [editingProduct, setEditingProduct] = useState<string | null>(null);
+  const [showProductForm, setShowProductForm] = useState(false);
+  const startEditProduct = (product: ProductWithUI) => {
+    setEditingProduct(product.id);
+    setProductForm({
+      name: product.name,
+      price: product.price,
+      stock: product.stock,
+      description: product.description || "",
+      discounts: product.discounts || [],
+    });
+    setShowProductForm(true);
+  };
+
+  return {
+    productForm,
+    setProductForm,
+    editingProduct,
+    setEditingProduct,
+    showProductForm,
+    setShowProductForm,
+    startEditProduct,
+  };
+};
